validate: refactored environments validation
closes #487
